some updates

+++ examples/QuadraticAssignment/heuristics.c
@@ -3,7 +3,7 @@
 #include "DataStructure.h"
 #include "utils.h"
 
-int assignmentCost(qa_global params, const int *assignment);
+long assignmentCost(qa_global params, const int *assignment);
 
 void swap(int *a, int *b);
 
@@ -26,7 +26,7 @@ stack_data* randomWithCorrectingIterations(qa_global params) {
             //swap factories
             swap(&heuristics->assignment[i], &heuristics->assignment[j]);
 
-            int cost = assignmentCost(params, heuristics->assignment);
+            long cost = assignmentCost(params, heuristics->assignment);
 
             if (cost > heuristics->cost) {
                 swap(&heuristics->assignment[j], &heuristics->assignment[i]);
@@ -39,8 +39,8 @@ stack_data* randomWithCorrectingIterations(qa_global params) {
     return heuristics;
 }
 
-int assignmentCost(qa_global params, const int *assignment) {
-    int cost = 0;
+long assignmentCost(qa_global params, const int *assignment) {
+    long cost = 0;
     for (int i = 0; i < params.locations; ++i) {
         for (int j = i; j < params.locations; ++j) {
             cost += params.distanceMatrix[i][j] * params.flowMatrix[assignment[i]][assignment[j]];
